refactor: pull most logic into helper functions/files
